pitsy-law/requirements.txt
openai==1.12.0
python-dotenv==1.0.1
streamlit==1.31.1
